+++ x/configuration/keeper.go
@@ -84,5 +84,5 @@ func (k Keeper) SetConfig(ctx sdk.Context, conf types.Config) {
 // GetDomainGracePeriod returns the default grace period before domains
 // can be deleted by someone other than the owner him/herself
 func (k Keeper) GetDomainGracePeriod(ctx sdk.Context) time.Duration {
-	return time.Duration(k.GetConfiguration(ctx).DomainGracePeriod) * time.Second
+	return k.GetConfiguration(ctx).DomainGracePeriod
 }