+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## HEAD
 
+- domain grace period is time.duration now 
 - refactor configuration module to be used with multisig wallets
 - allow empty account name on msg.Validate()
 - enable fees for all domain module handlers